Adding Files
This commit is relative to the authentication system of primeMovies - part 1, until here, it's only possible to run the main server and i'm creating the  navigation system of API (The router in this case), where i've created the two main routes, the sign-in and sign-up routes. But the isAuthenticated middleware isn't still builded and the functionalities to the routes.