Adding and Updating Files
In this commit, i've finished the implementation of middleware where i redone the request data type from middleware that checks if it is valid. I done the functionalities that changes the user informations of your own account and almost there to finish the controller class with the methods that  have the bussiness logical.