Updating Files
In this commit, i finished the functionalities that changes user credentials in the database and i've fixed some bugs where i wrong writed because i got confused with so much code and i was tired. Sorry Guys! In the future i'll implement more security layers and new strategies relative to the access token like JWT.